style fixes update tests
